adjustments for mps

+++ utils/math_utils.py
@@ -32,7 +32,7 @@ class Artanh(torch.autograd.Function):
     def forward(ctx, x):
         x = x.clamp(-1 + 1e-15, 1 - 1e-15)
         ctx.save_for_backward(x)
-        z = x.double()
+        z = x.to(torch.float32)
         return (torch.log_(1 + z).sub_(torch.log_(1 - z))).mul_(0.5).to(x.dtype)
 
     @staticmethod
@@ -45,7 +45,7 @@ class Arsinh(torch.autograd.Function):
     @staticmethod
     def forward(ctx, x):
         ctx.save_for_backward(x)
-        z = x.double()
+        z = x.to(torch.float32)
         return (z + torch.sqrt_(1 + z.pow(2))).clamp_min_(1e-15).log_().to(x.dtype)
 
     @staticmethod
@@ -59,7 +59,7 @@ class Arcosh(torch.autograd.Function):
     def forward(ctx, x):
         x = x.clamp(min=1.0 + 1e-15)
         ctx.save_for_backward(x)
-        z = x.double()
+        z = x.to(torch.float32)
         return (z + torch.sqrt_(z.pow(2) - 1)).clamp_min_(1e-15).log_().to(x.dtype)
 
     @staticmethod
